requirements.txt
# Night Watchman Bot Dependencies

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0
